docker/streamlit_demo/requirements.txt
streamlit==1.35.0
transformers==4.41.1
torch==2.3.0
st-annotated-text==4.0.1